tests/wpt/tests/workers/Worker_ErrorEvent_message.htm

<!DOCTYPE html>
<title> AbstractWorker ErrorEvent.message </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
// The worker events races with the window's load event; if the worker events
// arrive first, the harness will detect the error event and fail the test.
setup({ allow_uncaught_exception: true });

async_test(function(t) {
  var message = 'Error Message';
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onerror = t.step_func_done(function(e) {
    assert_greater_than(e.message.indexOf(message), -1);
  });
  worker.postMessage(message);
});
</script>
